+++ hutool-core/src/main/java/cn/hutool/core/util/StrUtil.java
@@ -242,6 +242,46 @@ public class StrUtil {
 	public static String nullToDefault(CharSequence str, String defaultStr) {
 		return (str == null) ? defaultStr : str.toString();
 	}
+	
+	/**
+	 * 如果字符串是<code>null</code>或者&quot;&quot;，则返回指定默认字符串，否则返回字符串本身。
+	 * 
+	 * <pre>
+	 * emptyToDefault(null, &quot;default&quot;)  = &quot;default&quot;
+	 * emptyToDefault(&quot;&quot;, &quot;default&quot;)    = &quot;default&quot;
+	 * emptyToDefault(&quot;  &quot;, &quot;default&quot;)  = &quot;  &quot;
+	 * emptyToDefault(&quot;bat&quot;, &quot;default&quot;) = &quot;bat&quot;
+	 * </pre>
+	 * 
+	 * @param str 要转换的字符串
+	 * @param defaultStr 默认字符串
+	 * 
+	 * @return 字符串本身或指定的默认字符串
+	 * @since 4.1.0
+	 */
+	public static String emptyToDefault(CharSequence str, String defaultStr) {
+		return isEmpty(str) ? defaultStr : str.toString();
+	}
+	
+	/**
+	 * 如果字符串是<code>null</code>或者&quot;&quot;或者空白，则返回指定默认字符串，否则返回字符串本身。
+	 * 
+	 * <pre>
+	 * emptyToDefault(null, &quot;default&quot;)  = &quot;default&quot;
+	 * emptyToDefault(&quot;&quot;, &quot;default&quot;)    = &quot;default&quot;
+	 * emptyToDefault(&quot;  &quot;, &quot;default&quot;)  = &quot;default&quot;
+	 * emptyToDefault(&quot;bat&quot;, &quot;default&quot;) = &quot;bat&quot;
+	 * </pre>
+	 * 
+	 * @param str 要转换的字符串
+	 * @param defaultStr 默认字符串
+	 * 
+	 * @return 字符串本身或指定的默认字符串
+	 * @since 4.1.0
+	 */
+	public static String blankToDefault(CharSequence str, String defaultStr) {
+		return isBlank(str) ? defaultStr : str.toString();
+	}
 
 	/**
 	 * 当给定字符串为空字符串时，转换为<code>null</code>