+++ hutool-extra/src/main/java/cn/hutool/extra/servlet/ServletUtil.java
@@ -209,7 +209,7 @@ public class ServletUtil {
 	 * @since 4.0.2
 	 */
 	public static MultipartFormData getMultipart(ServletRequest request) throws IORuntimeException {
-		return getMultipart(request);
+		return getMultipart(request, new UploadSetting());
 	}
 
 	/**